+++ packages/charts/react-charts/library/src/components/DeclarativeChart/DeclarativeChart.tsx
@@ -78,7 +78,6 @@ const ResponsiveGanttChart = withResponsiveContainer(GanttChart);
 
 // Default x-axis key for grouping traces. Also applicable for PieData and SankeyData where x-axis is not defined.
 const DEFAULT_XAXIS = 'x';
-const FALLBACK_TYPE = 'fallback';
 
 /**
  * DeclarativeChart schema.
@@ -516,9 +515,7 @@ export const DeclarativeChart: React.FunctionComponent<DeclarativeChartProps> = 
 
           const filteredTracesInfo = validTracesFilteredIndex.filter(trace => index.includes(trace.index));
           let chartType =
-            validTracesFilteredIndex.some(trace => trace.type === FALLBACK_TYPE) || chart.type === FALLBACK_TYPE
-              ? FALLBACK_TYPE
-              : filteredTracesInfo[0].type;
+            chart.type === 'fallback' || chart.type === 'groupedverticalbar' ? chart.type : filteredTracesInfo[0].type;
 
           if (
             validTracesFilteredIndex.some(trace => trace.type === 'line') &&