+++ packages/charts/react-charts/library/src/utilities/utilities.ts
@@ -260,6 +260,7 @@ export function createNumericXAxis(
     calcMaxLabelWidth,
     tickStep,
     tick0,
+    tickText,
   } = xAxisParams;
   const xAxisScale = createNumericScale(scaleType)
     .domain([domainNRangeValues.dStartValue, domainNRangeValues.dEndValue])
@@ -268,6 +269,9 @@ export function createNumericXAxis(
 
   let tickCount = xAxisCount ?? 6;
   const tickFormat = (domainValue: NumberValue, _index: number, defaultFormat?: (val: NumberValue) => string) => {
+    if (tickParams.tickValues && tickText && typeof tickText[_index] !== 'undefined') {
+      return tickText[_index];
+    }
     if (tickParams.tickFormat) {
       return d3Format(tickParams.tickFormat)(domainValue);
     }
@@ -437,6 +441,7 @@ export function createDateXAxis(
     calcMaxLabelWidth,
     tickStep,
     tick0,
+    tickText,
   } = xAxisParams;
   const isUtcSet = useUTC === true || useUTC === 'utc';
   const xAxisScale = isUtcSet ? d3ScaleUtc() : d3ScaleTime();
@@ -473,6 +478,9 @@ export function createDateXAxis(
   );
 
   const tickFormat = (domainValue: Date, _index: number) => {
+    if (tickParams.tickValues && tickText && typeof tickText[_index] !== 'undefined') {
+      return tickText[_index];
+    }
     if (customDateTimeFormatter) {
       return customDateTimeFormatter(domainValue);
     }
@@ -549,6 +557,7 @@ export function createStringXAxis(
     containerWidth,
     hideTickOverlap,
     calcMaxLabelWidth,
+    tickText,
   } = xAxisParams;
   const xAxisScale = d3ScaleBand()
     .domain(dataset!)
@@ -558,7 +567,10 @@ export function createStringXAxis(
 
   let tickValues = (tickParams.tickValues as string[] | undefined) ?? dataset;
   const tickFormat = (domainValue: string, _index: number) => {
-    return formatToLocaleString(domainValue, culture) as string;
+    if (tickParams.tickValues && tickText && typeof tickText[_index] !== 'undefined') {
+      return tickText[_index];
+    }
+    return domainValue;
   };
   if (hideTickOverlap) {
     let nonOverlappingTickValues = [];
@@ -708,6 +720,7 @@ export function createYAxisForHorizontalBarChartWithAxis(
     tickValues,
     tickStep,
     tick0,
+    tickText,
   } = yAxisParams;
 
   // maxOfYVal coming from horizontal bar chart with axis (Calculation done at base file)
@@ -719,7 +732,20 @@ export function createYAxisForHorizontalBarChartWithAxis(
     .range([containerHeight - margins.bottom!, margins.top!]);
   const axis = isRtl ? d3AxisRight(yAxisScale) : d3AxisLeft(yAxisScale);
   const yAxis = axis.tickPadding(tickPadding).ticks(yAxisTickCount);
-  yAxisTickFormat ? yAxis.tickFormat(yAxisTickFormat) : yAxis.tickFormat(defaultYAxisTickFormatter);
+  const tickFormat = (domainValue: NumberValue, index: number) => {
+    if (tickValues && tickText && typeof tickText[index] !== 'undefined') {
+      return tickText[index];
+    }
+    if (typeof yAxisTickFormat === 'function') {
+      return yAxisTickFormat(domainValue, index);
+    }
+    if (typeof yAxisTickFormat === 'string') {
+      return d3Format(yAxisTickFormat)(domainValue);
+    }
+    const value = typeof domainValue === 'number' ? domainValue : domainValue.valueOf();
+    return defaultYAxisTickFormatter(value);
+  };
+  yAxis.tickFormat(tickFormat);
 
   let customTickValues: number[] | undefined;
   if (tickValues) {
@@ -763,6 +789,7 @@ export function createNumericYAxis(
     tickValues,
     tickStep,
     tick0,
+    tickText,
   } = yAxisParams;
 
   // maxOfYVal coming from only area chart and Grouped vertical bar chart(Calculation done at base file)
@@ -816,12 +843,19 @@ export function createNumericYAxis(
   }
 
   const tickFormat = (domainValue: NumberValue, index: number, defaultFormat?: (val: NumberValue) => string) => {
+    if (tickValues && tickText && typeof tickText[index] !== 'undefined') {
+      return tickText[index];
+    }
+    if (typeof yAxisTickFormat === 'function') {
+      return yAxisTickFormat(domainValue, index);
+    }
+    if (typeof yAxisTickFormat === 'string') {
+      return d3Format(yAxisTickFormat)(domainValue);
+    }
     const value = typeof domainValue === 'number' ? domainValue : domainValue.valueOf();
     return defaultFormat?.(value) === '' ? '' : defaultYAxisTickFormatter(value);
   };
-  yAxisTickFormat
-    ? yAxis.tickFormat(yAxisTickFormat)
-    : yAxis.tickFormat((v, i) => tickFormat(v as NumberValue, i, yAxisScale.tickFormat(yAxisTickCount)));
+  yAxis.tickFormat((v, i) => tickFormat(v as NumberValue, i, yAxisScale.tickFormat(yAxisTickCount)));
   yAxisElement
     ? d3Select(yAxisElement)
         .call(yAxis)
@@ -846,7 +880,16 @@ export const createStringYAxisForHorizontalBarChartWithAxis = (
   isRtl: boolean,
   barWidth: number,
 ): ScaleBand<string> => {
-  const { containerHeight, tickPadding = 12, margins, yAxisTickFormat, yAxisElement, yAxisPadding } = yAxisParams;
+  const {
+    containerHeight,
+    tickPadding = 12,
+    margins,
+    yAxisTickFormat,
+    yAxisElement,
+    yAxisPadding,
+    tickValues,
+    tickText,
+  } = yAxisParams;
 
   let yAxisPaddingValue = yAxisPadding ?? 0.5;
   yAxisPaddingValue = yAxisPaddingValue === 1 ? 0.99 : yAxisPaddingValue;
@@ -855,10 +898,17 @@ export const createStringYAxisForHorizontalBarChartWithAxis = (
     .range([containerHeight - margins.bottom!, margins.top!])
     .padding(yAxisPaddingValue);
   const axis = isRtl ? d3AxisRight(yAxisScale) : d3AxisLeft(yAxisScale);
-  const yAxis = axis.tickPadding(tickPadding).ticks(dataPoints);
-  if (yAxisTickFormat) {
-    yAxis.tickFormat(yAxisTickFormat);
-  }
+  const customTickValues = (tickValues as string[] | undefined) ?? dataPoints;
+  const tickFormat = (domainValue: string, _index: number) => {
+    if (tickValues && tickText && typeof tickText[_index] !== 'undefined') {
+      return tickText[_index];
+    }
+    if (typeof yAxisTickFormat === 'function') {
+      return yAxisTickFormat(domainValue, _index);
+    }
+    return domainValue;
+  };
+  const yAxis = axis.tickPadding(tickPadding).tickValues(customTickValues).tickFormat(tickFormat);
   yAxisElement ? d3Select(yAxisElement).call(yAxis).selectAll('text') : '';
   return yAxisScale;
 };
@@ -884,6 +934,8 @@ export const createStringYAxis = (
     yAxisElement,
     yAxisPadding = 0,
     containerWidth,
+    tickValues,
+    tickText,
   } = yAxisParams;
   const yAxisScale = d3ScaleBand()
     .domain(dataPoints)
@@ -893,12 +945,19 @@ export const createStringYAxis = (
     yAxisScale.paddingInner(1).paddingOuter(0);
   }
   const axis = isRtl ? d3AxisRight(yAxisScale) : d3AxisLeft(yAxisScale);
-  const yAxis = axis.tickPadding(tickPadding).tickValues(dataPoints).tickSize(0);
+  const customTickValues = (tickValues as string[] | undefined) ?? dataPoints;
+  const tickFormat = (domainValue: string, _index: number) => {
+    if (tickValues && tickText && typeof tickText[_index] !== 'undefined') {
+      return tickText[_index];
+    }
+    if (typeof yAxisTickFormat === 'function') {
+      return yAxisTickFormat(domainValue, _index);
+    }
+    return domainValue;
+  };
+  const yAxis = axis.tickPadding(tickPadding).tickValues(customTickValues).tickFormat(tickFormat).tickSize(0);
   if (chartType === ChartTypes.VerticalStackedBarChart) {
     axis.tickSizeInner(-(containerWidth - margins.left! - margins.right!));
-  }
-  if (yAxisTickFormat) {
-    yAxis.tickFormat(yAxisTickFormat);
   }
   yAxisElement ? d3Select(yAxisElement).call(yAxis).selectAll('text') : '';
   return yAxisScale;