+++ packages/react/src/components/OverflowSet/OverflowSet.test.tsx
@@ -1,8 +1,6 @@
 import * as React from 'react';
 import * as path from 'path';
-import { RenderResult, act, fireEvent, render } from '@testing-library/react';
-import * as renderer from 'react-test-renderer';
-import { create } from '@fluentui/test-utilities';
+import { type RenderResult, act, fireEvent, render } from '@testing-library/react';
 
 import { CommandBarButton } from '../../Button';
 import { KeytipLayer, KeytipLayerBase } from '../../KeytipLayer';
@@ -45,23 +43,25 @@ const runAllTimers = () =>
 describe('OverflowSet', () => {
   describe('snapshot tests', () => {
     test('basicSnapshot', () => {
-      const component = create(<OverflowSet onRenderItem={noOp} onRenderOverflowButton={noOp} />);
-      const tree = component.toJSON();
-      expect(tree).toMatchSnapshot();
+      const { container } = render(<OverflowSet onRenderItem={noOp} onRenderOverflowButton={noOp} />);
+
+      expect(container.firstChild).toMatchSnapshot();
     });
 
     test('snapshot with classname', () => {
-      const component = create(<OverflowSet className="foobar" onRenderItem={noOp} onRenderOverflowButton={noOp} />);
-      const tree = component.toJSON();
-      expect(tree).toMatchSnapshot();
+      const { container } = render(
+        <OverflowSet className="foobar" onRenderItem={noOp} onRenderOverflowButton={noOp} />,
+      );
+
+      expect(container.firstChild).toMatchSnapshot();
     });
 
     test('snapshot with classname and vertical layout', () => {
-      const component = create(
+      const { container } = render(
         <OverflowSet className="foobar" vertical onRenderItem={noOp} onRenderOverflowButton={noOp} />,
       );
-      const tree = component.toJSON();
-      expect(tree).toMatchSnapshot();
+
+      expect(container.firstChild).toMatchSnapshot();
     });
   });
 
@@ -73,16 +73,14 @@ describe('OverflowSet', () => {
 
   it('does not render overflow when there are no overflow items', () => {
     const onRenderOverflowButton = jest.fn();
-    renderer.create(<OverflowSet onRenderItem={noOp} onRenderOverflowButton={onRenderOverflowButton} />);
+    render(<OverflowSet onRenderItem={noOp} onRenderOverflowButton={onRenderOverflowButton} />);
 
     expect(onRenderOverflowButton).not.toHaveBeenCalled();
   });
 
   it('does not render overflow when overflow items is an empty array', () => {
     const onRenderOverflowButton = jest.fn();
-    renderer.create(
-      <OverflowSet onRenderItem={noOp} onRenderOverflowButton={onRenderOverflowButton} overflowItems={[]} />,
-    );
+    render(<OverflowSet onRenderItem={noOp} onRenderOverflowButton={onRenderOverflowButton} overflowItems={[]} />);
 
     expect(onRenderOverflowButton).not.toHaveBeenCalled();
   });