+++ packages/react-icon-provider/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-icon-provider
 
-This log was last generated on Tue, 14 Oct 2025 07:21:45 GMT and should not be manually modified.
+This log was last generated on Thu, 16 Oct 2025 07:22:01 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.3.88](https://github.com/microsoft/fluentui/tree/@fluentui/react-icon-provider_v1.3.88)
+
+Thu, 16 Oct 2025 07:22:01 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-icon-provider_v1.3.87..@fluentui/react-icon-provider_v1.3.88)
+
+### Patches
+
+- Bump @fluentui/style-utilities to v8.13.2 ([commit](https://github.com/microsoft/fluentui/commit/36be47512933fadd0418d11cdc1e3530c13f0575) by beachball)
 
 ## [1.3.87](https://github.com/microsoft/fluentui/tree/@fluentui/react-icon-provider_v1.3.87)